battery recharge icons, move energy bar
I've made some improvements to the UI. I wanted to make the battery draining/recharge more apparent to the user. So I move the battery bar to the right side of the screen, and also put a battery icon that changes based on the stamina to represent about how much power you have.